map di un oggetto per i moduli
