+++ frontend/src/App.tsx
@@ -1,6 +1,6 @@
 import { Routes, Route } from "react-router-dom"
 import AccessPage from "./pages/AccessPage.tsx"
-import { Homepage } from "./Homepage.tsx"
+import { Homepage } from "./pages/Homepage.tsx"
 import UserPage from "./pages/UserPage.tsx"
 import Quiz from "./pages/Quiz.tsx"
 import EndModule from "./pages/end-module/EndModule.tsx"
