+++ frontend/src/moduli.json
@@ -1,0 +1,50 @@
+[
+	{
+		"argomento": "html",
+		"img": "../public/icone/html.jpg",
+		"check": "../public/icone/check.jpg",
+		"value": 100
+	},
+	{
+		"argomento": "css",
+		"img": "../public/icone/css.jpg",
+		"check": "../public/icone/open-lock.png",
+		"value": 70
+	},
+	{
+		"argomento": "javascript",
+		"img": "../public/icone/js.png",
+		"check": "../public/icone/closed-lock.png",
+		"value": 100
+	},
+	{
+		"argomento": "typescript",
+		"img": "../public/icone/ts.png",
+		"check": "../public/icone/closed-lock.png",
+		"value": 100
+	},
+	{
+		"argomento": "react",
+		"img": "../public/icone/react.png",
+		"check": "../public/icone/closed-lock.png",
+		"value": 100
+	},
+	{
+		"argomento": "sql",
+		"img": "../public/icone/sql.png",
+		"check": "../public/icone/closed-lock.png",
+		"value": 100
+	},
+	{
+		"argomento": "git",
+		"img": "../public/icone/git.png",
+		"check": "../public/icone/closed-lock.png",
+		"value": 100
+	},
+	{
+		"argomento": "node",
+		"img": "../public/icone/nodejs.png",
+		"check": "../public/icone/closed-lock.png",
+		"value": 100
+	}
+]
